requirements.txt
streamlit>=1.28.0
tensorflow-macos>=2.15.0
tensorflow-metal>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
